delete verification alert

+++ store/view/orders.php
@@ -16,7 +16,7 @@ foreach($this->model->getordersArray() as $orders){
             $str.="<td>".$orders->getStatus()."</td>";
             $str.="<td>".$orders->getCreatedtime()."</td>";
             $str.="<td><a id='button' href='clientproducts.php?userid=".$orders->getUserid()."&orderid=".$orders->getId()."'>View</a></td>";
-            $str.="<td><a id='button' href='orders.php?action=delete&orderid=".$orders->getId()."'>Delete</a></td>";
+            $str.="<td><a id='button' class='Delete' onclick='return confirm('Delete this order?');' href='orders.php?action=delete&orderid=".$orders->getId()."'>Delete</a></td>";
             $str.="<tr>";
           }
             $i++;
